refactor(grammar): extract concept roster to shared module, break shared→platform dependency
Move GRAMMAR_MONSTER_CONCEPTS, GRAMMAR_AGGREGATE_CONCEPTS, and
GRAMMAR_CONCEPT_TO_MONSTER from src/platform/game/mastery/grammar.js to
shared/grammar/grammar-concept-roster.js. The platform module re-exports
them for backward compatibility. shared/grammar/grammar-stars.js now
imports from the shared roster instead of from the platform layer,
eliminating the shared→platform dependency direction violation.

Also fix two pre-existing P6-U6 test failures: add createdAt timestamps
to recentAttempts fixtures so retainedAfterSecure temporal proof passes.

Phase 7 U1. No behavioural change. All 267 Star tests + 116 UI model
tests pass. Client bundle audit passes.
